separate edit and new game
