+++ src/components/templates/Layout/index.tsx
@@ -8,14 +8,15 @@ import styles from './Layout.module.scss'
 interface Props extends HeaderProps{
   children: React.ReactNode;
   buttons?: ButtonProps[];
+  loading?: boolean;
 }
 
-const Layout = ({ title, subTitle, letter, onClose, children, buttons }: Props) => {
+const Layout = ({ title, subTitle, letter, onClose, children, buttons, loading }: Props) => {
   return (
     <div className={styles.container}>
       <Header title={title} subTitle={subTitle} onClose={onClose} letter={letter}/>
       <main className={styles.main}>
-        {children}
+        {loading ? 'Cargando...' : children}
       </main>
       {buttons && (
         <footer className={styles.footer}>
