+++ src/pages/NewGame/index.tsx
@@ -1,14 +1,15 @@
 import React, { useContext } from 'react'
 import { useHistory } from 'react-router'
-import GameConfig from '../../components/Organisms/GameConfig'
+import GameConfigForm from '../../components/Organisms/GameConfigForm'
 import { GAME_CONFIG_FIELDS } from '../../components/Organisms/GameConfigForm/constants'
+import Layout from '../../components/templates/Layout'
 import { PlayerContext } from '../../contexts/PlayerContextState'
 import { addRoom } from '../../firebase/services/room'
 import { iCategory } from '../../hooks/useCategories/types'
 import { useForm } from '../../hooks/useForm'
 import { paths } from '../../routes'
 import { IRoom } from '../../types/room'
-import { INITIAL_STATE_GAME } from './constants'
+import { FOOTER_BUTTONS, INITIAL_STATE_GAME } from './constants'
 
 const NewGame = () => {
   const [values, handleChange,, setValue] = useForm<IRoom>(INITIAL_STATE_GAME)
@@ -29,14 +30,20 @@ const NewGame = () => {
   }
 
   return (
-    <GameConfig
-      values={values}
-      handleChange={handleChange}
-      setValue={setValue}
-      categories={values.categories}
-      setCategories={handleChangeCategories}
-      handleSubmit={handleSubmit}
-    />
+    <Layout
+      title="Creación de partida"
+      subTitle="Configuración de parámetros"
+      onClose={() => history.push(paths.HOME)}
+      buttons={FOOTER_BUTTONS(handleSubmit)}
+    >
+      <GameConfigForm
+        values={values}
+        handleChange={handleChange}
+        setValue={setValue}
+        categories={values.categories}
+        setCategories={handleChangeCategories}
+      />
+    </Layout>
   )
 }
